+++ permissions/change/revokePermission.ps1
@@ -263,12 +263,12 @@ function Confirm-Description {
         $AllowedLength
     )
     if ($Description.Length -gt $AllowedLength) {
-        $errorMessage = "Could not revoke TOPdesk entitlement [$id]: The attribute [$AttributeName] exceeds the max amount of [$AllowedLength] characters. Please shorten the value for this attribute in the JSON file. Value: [$Description]"
-        
-        $outputContext.AuditLogs.Add([PSCustomObject]@{
-                Message = $errorMessage
-                IsError = $true
-            })
+        Write-Information "Attribute [$AttributeName] exceeds [$AllowedLength] characters [$Description] and will be shortened."
+        $descriptionShortened = $Description.substring(0, [System.Math]::Min($AllowedLength, $Description.Length))
+        return $descriptionShortened
+    }
+    else {
+        return $Description
     }
 }
 
@@ -733,13 +733,12 @@ try {
         Description   = $briefDescription
         AllowedLength = 80
         AttributeName = 'BriefDescription'
-        id            = $pref.id
+        id            = $pRef.id
     }
-    Confirm-Description @splatParamsValidateBriefDescription
-
+    
     # Add value to request object
     $requestObject += @{
-        briefDescription = $briefDescription
+        briefDescription = Confirm-Description @splatParamsValidateBriefDescription
     }
 
     # Resolve variables in the request field