+++ permissions/incident/revokePermission.ps1
@@ -226,12 +226,12 @@ function Confirm-Description {
         $AllowedLength
     )
     if ($Description.Length -gt $AllowedLength) {
-        $errorMessage = "Could not revoke TOPdesk entitlement [$id]: The attribute [$AttributeName] exceeds the max amount of [$AllowedLength] characters. Please shorten the value for this attribute in the JSON file. Value: [$Description]"
-        
-        $outputContext.AuditLogs.Add([PSCustomObject]@{
-                Message = $errorMessage
-                IsError = $true
-            })
+        Write-Information "Attribute [$AttributeName] exceeds [$AllowedLength] characters [$Description] and will be shortened."
+        $descriptionShortened = $Description.substring(0, [System.Math]::Min($AllowedLength, $Description.Length))
+        return $descriptionShortened
+    }
+    else {
+        return $Description
     }
 }
 
@@ -525,6 +525,11 @@ function Get-TopdeskIdentifier {
 
     $result = $responseGet | Where-object $SearchAttribute -eq $Value
 
+    if ($class -eq 'SubCategory' -and ($result | Measure-Object).Count -gt 1) {
+        # Ensure category ID is available or adjust order of operations to guarantee its presence
+        $result = $result | Where-Object { $_.Category.Id -eq $RequestObject.category.id }
+    }
+
     # When attribute $Class with $Value is not found in Topdesk
     if ([string]::IsNullOrEmpty($result.id)) {
         $errorMessage = "Class [$Class] with SearchAttribute [$SearchAttribute] with value [$Value] isn't found in Topdesk"
@@ -740,14 +745,12 @@ try {
         Description   = $requestShort
         AllowedLength = 80
         AttributeName = 'requestShort'
-        id            = $pref.id
+        id            = $pRef.id
     }
-
-    Confirm-Description @splatParamsValidateRequestShort
     
     # Add value to request object
     $requestObject += @{
-        briefDescription = $requestShort
+        briefDescription = Confirm-Description @splatParamsValidateRequestShort
     }
 
     # Resolve variables in the RequestDescription field